refactor(api_jobs): decompose god-file test module into submodules
api_jobs/mod.rs was 2507 lines (god_file) but was essentially a module
re-export header (~20 lines) plus a ~2485-line inline test module — the
production api_jobs code already lives in submodules (persistence,
remote_runner, store, summary, types).

Split the test module into api_jobs/tests/{part_a,part_b,part_c}.rs
(3-way, by line count, since a 2-way split left one half over 1500) with
shared helpers (record_test_local_child, remote_runner_request) promoted
to api_jobs/tests/mod.rs as pub(super).

mod.rs is now 23 lines (re-exports + tests decl). cargo check -p
homeboy-core --lib is clean. Clears the god_file finding.
